# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: RDV.BI API
  version: 0.1.0
  summary: Appointment booking, queue, and institution administration API
  description: |
    OpenAPI contract for the current RDV.BI backend.

    Notes:
    - Public endpoints live under `/api/public`.
    - Authentication endpoints live under `/api/auth`.
    - Admin endpoints require a bearer JWT issued by `/api/auth/login`.
    - `POST /api/auth/bootstrap-admin` exists only for bootstrap and should be removed or disabled in production.
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh an access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /api/auth/bootstrap-admin:
    post:
      tags: [Auth]
      summary: Bootstrap the first institution admin
      description: Internal setup endpoint. Remove or disable in production.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapAdminRequest'
      responses:
        '200':
          description: Admin created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: created
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /api/public/institutions/register:
    post:
      tags: [Public Institutions]
      summary: Register an institution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstitutionRegistrationRequest'
      responses:
        '201':
          description: Institution registration submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstitutionRegistrationResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/public/institutions/{slug}:
    get:
      tags: [Public Institutions]
      summary: Get public institution details
      parameters:
        - $ref: '#/components/parameters/SlugParam'
        - name: date
          in: query
          required: false
          description: Filter time slots by date in `YYYY-MM-DD`
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Public institution payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicInstitutionPayload'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /api/public/appointments:
    post:
      tags: [Public Appointments]
      summary: Create an appointment booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentRequest'
      responses:
        '201':
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentCreateResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SlugParam:
      name: slug
      in: path
      required: true
      schema:
        type: string
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
      example:
        error: Institution not found
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      required: [email, password]
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required: [refresh_token]
    BootstrapAdminRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        fullName:
          type: string
        password:
          type: string
          format: password
        institutionId:
          type: string
          format: uuid
      required: [email, password, institutionId]
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/AuthUser'
      required: [accessToken, refreshToken, user]
    AuthUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        institutionId:
          type: string
          format: uuid
        institutionName:
          type: string
      required: [id, email, fullName, role, institutionId, institutionName]
    InstitutionRegistrationRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
        type:
          type: string
        contactEmail:
          type: string
          format: email
        contactPhone:
          type: string
        address:
          type: string
        description:
          type: string
        logoUrl:
          type: string
        locationLat:
          type: number
        locationLng:
          type: number
      required: [name, slug, contactPhone]
    InstitutionRegistrationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/InstitutionStatus'
      required: [id, status]
...
    AppointmentRequest:
      type: object
      properties:
        institutionId:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        timeSlotId:
          type: string
          format: uuid
        customerPhone:
          type: string
        customerEmail:
          type: string
          format: email
        appointmentDate:
          type: string
          format: date
        appointmentTime:
          type: string
          format: time
        ticketNumber:
          type: string
      required:
        - institutionId
        - serviceId
        - timeSlotId
        - customerPhone
        - appointmentDate
        - appointmentTime
        - ticketNumber
...
    AppointmentCreateResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/AppointmentStatus'
      required: [id, status]
    PublicInstitutionPayload:
      type: object
      properties:
        institution:
          $ref: '#/components/schemas/Institution'
        services:
          type: array
          items:
            $ref: '#/components/schemas/Service'
        timeSlots:
          type: array
          items:
            $ref: '#/components/schemas/TimeSlot'
      required: [institution, services, timeSlots]
    Institution:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        type:
          type: string
        contactEmail:
          type: string
          format: email
        contactPhone:
          type: string
        address:
          type: string
        description:
          type: string
        logoUrl:
          type: string
        locationLat:
          type: number
        locationLng:
          type: number
        status:
          $ref: '#/components/schemas/InstitutionStatus'
        active:
          type: boolean
        approvalNotes:
          type: string
        approvedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, slug, status, active]
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        prefix:
          type: string
        active:
          type: boolean
      required: [id, name, durationMinutes, active]
    TimeSlot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
        capacity:
          type: integer
        bookedCount:
          type: integer
        blocked:
          type: boolean
        blockReason:
          type: string
      required: [id, date, startTime, endTime, capacity, bookedCount, blocked]
...
    InstitutionStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
...
    AppointmentStatus:
      type: string
      enum: [PENDING, CONFIRMED, WAITING, IN_PROGRESS, COMPLETED, CANCELLED, NO_SHOW]
    UserRole:
      type: string
      enum: [ADMIN, STAFF]